refactor: simplify review code by co-locating functions with consumers
- Inline `displayReviewRoleTitle` logic directly in ReviewActionButton.vue
- Move `allowUserToApplyReviewAction` to ReviewActionButton.vue and simplify by combining early return conditions
- Remove unused `provideIssueReviewContext` from CICDLayout.vue
- Delete 3 files containing only single-use or unused code:
  - frontend/src/components/IssueV1/logic/review.ts
  - frontend/src/types/review.ts
  - frontend/src/components/Plan/logic/issue-review.ts
- Update index.ts exports accordingly

Net reduction: 94 lines
