+++ frontend/src/components/IssueV1/logic/index.ts
@@ -1,7 +1,6 @@
 export * from "./context";
 export * from "./base";
 export * from "./releaser";
-export * from "./review";
 export * from "./rollout";
 export * from "./utils";
 export * from "./initialize";
