+++ frontend/src/components/Plan/logic/index.ts
@@ -6,5 +6,4 @@ export * from "./sheet";
 export * from "./plan-check";
 export * from "./utils";
 export * from "./sidebar";
-export * from "./issue-review";
 export * from "./usePlanCheckStatus";
